+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## UNRELEASED
+
+- feat: add readfrom json tag to support reverse edges
+  [#49](https://github.com/hypermodeinc/modusDB/pull/49)
+
 ## 2025-01-02 - Version 0.1.0
 
 Baseline for the changelog.